+++ auth/providers/eks/kubectl.go
@@ -1,4 +1,4 @@
-package aws
+package eks
 
 import (
 	"encoding/json"
